controller 中止処理追加

+++ app/Http/Controllers/Api/Instructor/TagController.php
@@ -122,8 +122,14 @@ class TagController extends Controller
         $user = Auth::guard('instructor')->user();
         $tag = Tag::findOrFail($request->tag_id);
 
+        // タグの所有者が現在のログイン講師でない場合は処理を中断（認可エラー）
         if ($user->id !== $tag->instructor_id) {
             throw new AuthorizationException('Forbidden, invalid instructor.');
+        }
+
+        // タグに紐づく講座が存在する場合は削除処理を中止
+        if ($tag->courses()->exists()) {
+            throw new AuthorizationException('Forbidden, this tag is linked to courses.');
         }
 
         $tag->delete();
